+++ frontend/messages/en.json
@@ -1225,7 +1225,7 @@
 	"potentialThreats": "potential threats",
 	"followUpUnresolvedHigh": "Unresolved High or Critical",
 	"level": "Level",
-	"quickStart": "🚀 Quick Start",
+	"quickStart": "Quick Start",
 	"auditName": "Audit Name",
 	"createRiskAssessment": "Initiate a risk assessment as well",
 	"riskAssessmentName": "Risk Assessment Name"
