+++ frontend/messages/fr.json
@@ -1213,7 +1213,7 @@
 	"potentialThreats": "menaces potentielles",
 	"followUpUnresolvedHigh": "Hauts ou critiques non résolus",
 	"level": "Niveau",
-	"quickStart": "🚀 Démarrage Rapide",
+	"quickStart": "Démarrage Rapide",
 	"auditName": "Nom de l'audit",
 	"createRiskAssessment": "Initier une analyse de risque également",
 	"riskAssessmentName": "Nom de l'analyse de risque"
